norme raycasting
